+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-23T11:44:41.290Z")
+    created: Optional[datetime] = parse_datetime("2025-10-23T13:46:15.290Z")
     r"""The date and time the resource was created"""