+++ packages/orq-rc/src/orq_ai_sdk/models/listbudgetsop.py
@@ -242,7 +242,7 @@ class ListBudgetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-23T11:44:38.219Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-23T13:46:12.112Z")
     r"""The date and time the resource was last updated"""